requirements.txt
pandas>=1.5.0
openpyxl>=3.1.0  # Para suporte a arquivos Excel
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0  # Driver PostgreSQL
nest_asyncio
opencv-python
numpy
pytesseract
playwright
